+++ src/utils/healthCheck.ts
@@ -3,7 +3,6 @@ import {
   ErgoNodeAssetHealthCheckParam,
 } from '@rosen-bridge/asset-check';
 import { HealthCheck, HealthStatusLevel } from '@rosen-bridge/health-check';
-import { ErgoNodeSyncHealthCheckParam } from '@rosen-bridge/node-sync-check';
 import {
   AbstractPermitHealthCheckParam,
   ExplorerPermitHealthCheckParam,
@@ -87,14 +86,6 @@ class HealthCheckSingleton {
       };
     }
     this.healthCheck = new HealthCheck(notify, notificationConfig);
-    const ergoNodeSyncCheck = new ErgoNodeSyncHealthCheckParam(
-      getConfig().healthCheck.ergoNodeMaxHeightDiff,
-      getConfig().healthCheck.ergoNodeMaxBlockTime,
-      getConfig().healthCheck.ergoNodeMinPeerCount,
-      getConfig().healthCheck.ergoNodeMaxPeerHeightDifference,
-      getConfig().general.nodeUrl
-    );
-    this.healthCheck.register(ergoNodeSyncCheck);
 
     const warnLogCheck = new LogLevelHealthCheck(
       CallbackLoggerFactory.getInstance(),
@@ -156,9 +147,10 @@ class HealthCheckSingleton {
     this.ergoScannerSyncCheckParam = new ScannerSyncHealthCheckParam(
       ERGO_CHAIN_NAME,
       this.observingNetworkLastBlock(scanner.getErgoScanner().name()),
-      getConfig().healthCheck.ergoScannerWarnDiff,
-      getConfig().healthCheck.ergoScannerCriticalDiff,
-      ERGO_BLOCK_TIME
+      getConfig().healthCheck.scannerWarnDiff,
+      getConfig().healthCheck.scannerCriticalDiff,
+      ERGO_BLOCK_TIME,
+      getConfig().general.ergoInterval
     );
     this.healthCheck.register(this.ergoScannerSyncCheckParam);
   };
@@ -190,9 +182,10 @@ class HealthCheckSingleton {
     this.ergoScannerSyncCheckParam = new ScannerSyncHealthCheckParam(
       ERGO_CHAIN_NAME,
       this.observingNetworkLastBlock(scanner.getErgoScanner().name()),
-      getConfig().healthCheck.ergoScannerWarnDiff,
-      getConfig().healthCheck.ergoScannerCriticalDiff,
-      ERGO_BLOCK_TIME
+      getConfig().healthCheck.scannerWarnDiff,
+      getConfig().healthCheck.scannerCriticalDiff,
+      ERGO_BLOCK_TIME,
+      getConfig().general.ergoInterval
     );
     this.healthCheck.register(this.ergoScannerSyncCheckParam);
   };
@@ -215,47 +208,55 @@ class HealthCheckSingleton {
         (
           scanner.getObservationScanner() as CardanoOgmiosScanner
         ).getConnectionStatus,
-        getConfig().healthCheck.cardanoScannerWarnDiff,
-        getConfig().healthCheck.cardanoScannerCriticalDiff,
+        getConfig().healthCheck.scannerWarnDiff,
+        getConfig().healthCheck.scannerCriticalDiff,
         // TODO: Fix configuration: local/health-check/-/issues/29
         getConfig().cardano.ogmios!.connectionRetrialInterval * 15
       );
     } else {
       let chainName: string;
       let chainBlockTime: number;
+      let updateInterval: number;
       switch (getConfig().general.networkWatcher) {
         case CARDANO_CHAIN_NAME:
           chainName = CARDANO_CHAIN_NAME;
           chainBlockTime = CARDANO_BLOCK_TIME;
+          updateInterval = getConfig().cardano.koios!.interval;
           break;
         case BITCOIN_CHAIN_NAME:
           chainName = BITCOIN_CHAIN_NAME;
           chainBlockTime = BITCOIN_BLOCK_TIME;
+          updateInterval = getConfig().bitcoin.interval;
           break;
         case BITCOIN_RUNES_CHAIN_NAME:
           chainName = BITCOIN_RUNES_CHAIN_NAME;
           chainBlockTime = BITCOIN_BLOCK_TIME;
+          updateInterval = getConfig().bitcoin.interval;
           break;
         case DOGE_CHAIN_NAME:
           chainName = DOGE_CHAIN_NAME;
           chainBlockTime = DOGE_BLOCK_TIME;
+          updateInterval = getConfig().doge.interval;
           break;
         case ETHEREUM_CHAIN_NAME:
           chainName = ETHEREUM_CHAIN_NAME;
           chainBlockTime = ETHEREUM_BLOCK_TIME;
+          updateInterval = getConfig().ethereum.interval;
           break;
         case BINANCE_CHAIN_NAME:
           chainName = BINANCE_CHAIN_NAME;
           chainBlockTime = BINANCE_BLOCK_TIME;
+          updateInterval = getConfig().binance.interval;
           break;
       }
 
       scannerSyncCheck = new ScannerSyncHealthCheckParam(
         chainName!,
         this.observingNetworkLastBlock(scanner.getObservationScanner().name()),
-        getConfig().healthCheck.cardanoScannerWarnDiff,
-        getConfig().healthCheck.cardanoScannerCriticalDiff,
-        chainBlockTime!
+        getConfig().healthCheck.scannerWarnDiff,
+        getConfig().healthCheck.scannerCriticalDiff,
+        chainBlockTime!,
+        updateInterval!
       );
     }
     this.healthCheck.register(scannerSyncCheck);